Refactor MappedSchema
* move `Field` to separate file
* remove unused constructors
