+++ protostuff-runtime/src/main/java/io/protostuff/runtime/RuntimeCollectionField.java
@@ -24,7 +24,6 @@ import io.protostuff.Output;
 import io.protostuff.Pipe;
 import io.protostuff.Tag;
 import io.protostuff.WireFormat.FieldType;
-import io.protostuff.runtime.MappedSchema.Field;
 
 /**
  * A runtime field for a {@link Collection}. Null values are not written.
