+++ protostuff-runtime/src/main/java/io/protostuff/runtime/RuntimeDerivativeField.java
@@ -20,7 +20,6 @@ import io.protostuff.Input;
 import io.protostuff.Schema;
 import io.protostuff.Tag;
 import io.protostuff.WireFormat.FieldType;
-import io.protostuff.runtime.MappedSchema.Field;
 
 /**
  * A runtime field w/c represents an abstract class, interface or a base type with many possible subclasses.
