+++ protostuff-runtime/src/main/java/io/protostuff/runtime/RuntimeMapFieldFactory.java
@@ -30,7 +30,6 @@ import io.protostuff.Pipe;
 import io.protostuff.Schema;
 import io.protostuff.Tag;
 import io.protostuff.WireFormat.FieldType;
-import io.protostuff.runtime.MappedSchema.Field;
 
 /**
  * Static utility for creating runtime {@link java.util.Map} fields.
